CaPO4Sim

Explore calcium (Ca) and phosphate (Pi) homeostasis with two novel 'Shiny' apps, building upon on a previously published mathematical model written in C, to ensure efficient computations. The underlying model is accessible here <https://www.ncbi.nlm.nih.gov/pubmed/28747359>. The first application explores the fundamentals of Ca-Pi homeostasis, while the second provides interactive case studies for in-depth exploration of the topic, thereby seeking to foster student engagement and an integrative understanding of Ca-Pi regulation. These applications are hosted at <https://rinterface.com/AppsPhysiol.html>.

Author

Maintainer: David Granjon dgranjon@ymail.com [copyright holder]

Other contributors:

Diane de Zélicourt [copyright holder]

Vartan Kurtcuoglu [copyright holder]

Olivier Bonny [copyright holder]

François Verrey [copyright holder]

University of Lausanne [funder]

University of Zurich [funder]

Kidney NCCR.CH [funder]

The Interface Group (Hosting Group) [copyright holder]

RinteRface (R/HTML Templates) [copyright holder]