carousel container

Creates a carousel. Adapted from shinydashboardplus to also allow control of the carousel animation

Usage

myCarousel(
  ...,
  id,
  indicators = TRUE,
  width = 6,
  .list = NULL,
  data.interval = 5000,
  data.ride = "carousel"
)

Arguments

...

Slot for carouselItem

id

Carousel id. Must be unique.

indicators

Whether to display left and right indicators.

width

Carousel width. 6 by default.

.list

Should you need to pass carouselItem via lapply or similar, put these item here instead of passing them in ...

data.interval

specify data-interval in ms. 5000ms by default, set to "false" to prevent automated animation of the slides.

data.ride

specify data-ride. "carousel" by default.